refactor(cli): tree api with injectable logging
