+++ packages/aws-cdk/test/api/tree.test.ts
@@ -1,6 +1,6 @@
 /* eslint-disable import/order */
 import * as path from 'path';
-import { ConstructTreeNode, loadTreeFromDir, some } from '../lib/tree';
+import { ConstructTreeNode, loadTreeFromDir, some } from '../../lib/api/tree';
 
 describe('some', () => {
   const tree: ConstructTreeNode = {
@@ -104,12 +104,12 @@ describe('some', () => {
 
 describe('loadTreeFromDir', () => {
   test('can find tree', () => {
-    const tree = loadTreeFromDir(path.join(__dirname, 'cloud-assembly-trees', 'built-with-1_144_0'));
+    const tree = loadTreeFromDir(path.join(__dirname, '..', 'cloud-assembly-trees', 'built-with-1_144_0'), () => {});
     expect(tree?.id).toEqual('App');
   });
 
   test('cannot find tree', () => {
-    const tree = loadTreeFromDir(path.join(__dirname, 'cloud-assembly-trees', 'foo'));
+    const tree = loadTreeFromDir(path.join(__dirname, '..', 'cloud-assembly-trees', 'foo'), () => {});
     expect(tree).toEqual(undefined);
   });
 });
